+++ lib/src/ui/layout/inherited_questionnaire_renderer.dart
@@ -21,6 +21,7 @@ final class InheritedQuestionnaireRenderer extends InheritedWidget {
     this.boolItemBuilder,
     this.displayItemBuilder,
     this.referenceItemBuilder,
+    this.attachmentItemBuilder,
     this.readOnly = false,
   });
 
@@ -51,6 +52,8 @@ final class InheritedQuestionnaireRenderer extends InheritedWidget {
   final QuestionnaireChoiceWidgetBuilder? openChoiceItemBuilder;
 
   final QuestionnaireReferenceWidgetBuilder? referenceItemBuilder;
+
+  final QuestionnaireAttachmentWidgetBuilder? attachmentItemBuilder;
 
   final void Function(QuestionnaireResponse response) onResponseChanged;